DELIA-59395: [Coverity] Various medium impact issues identified in aamp - part 5
Reason for change: Coverity issues are addressed as requested from security team.

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_23_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_23_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_23_1

+++ isobmff/isobmffbox.cpp
@@ -424,7 +424,7 @@ MdhdBox* MdhdBox::constructMdhdBox(uint32_t sz, uint8_t *ptr)
 {
 	uint8_t version = READ_VERSION(ptr);
 	uint32_t flags  = READ_FLAGS(ptr);
-	uint64_t tScale;
+	uint32_t tScale;
 
 	uint32_t skip = sizeof(uint32_t)*2;
 	if (1 == version)
